+++ resources/views/pagination/livewire.blade.php
@@ -11,18 +11,18 @@
         <!-- Mobile Pagination -->
         <div class="hidden on-sm:flex space-between items-center">
             @if ($paginator->onFirstPage())
-                <button type="button" class="btn xs" disabled> Previous </button>
+                <button type="button" class="btn primary xs" disabled> Previous </button>
             @else
                 <button wire:click="previousPage('{{ $paginator->getPageName() }}')"
                     x-on:click="{{ $scrollIntoViewJsSnippet }}" wire:loading.attr="disabled"
-                    type="button" class="btn xs"> Previous </button>
+                    type="button" class="btn primary xs"> Previous </button>
             @endif
             @if ($paginator->hasMorePages())
                 <button wire:click="nextPage('{{ $paginator->getPageName() }}')"
                     x-on:click="{{ $scrollIntoViewJsSnippet }}" wire:loading.attr="disabled"
-                    type="button" class="btn xs"> Next </button>
+                    type="button" class="btn primary xs"> Next </button>
             @else
-                <button type="button" class="btn xs" disabled> Next </button>
+                <button type="button" class="btn primary xs" disabled> Next </button>
             @endif
         </div>
 
@@ -40,14 +40,14 @@
                 {{-- Previous Icon --}}
                 <div class="inline-flex">
                     @if ($paginator->onFirstPage())
-                        <button type="button" class="btn px-1 py-05" disabled aria-label="Previous page">
+                        <button type="button" class="btn wh-2.5" disabled aria-label="Previous page">
                             <svg class="icon txt-muted" fill="currentColor" viewBox="0 0 20 20">
                                 <path fill-rule="evenodd" d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z" clip-rule="evenodd" />
                             </svg>
                         </button>
                     @else
                         <button wire:click="previousPage('{{ $paginator->getPageName() }}')" x-on:click="{{ $scrollIntoViewJsSnippet }}"
-                            type="button" class="btn px-1 py-05" aria-label="Previous page">
+                            type="button" class="btn wh-2.5" aria-label="Previous page">
                             <svg class="icon" fill="currentColor" viewBox="0 0 20 20">
                                 <path fill-rule="evenodd" d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z" clip-rule="evenodd" />
                             </svg>
@@ -60,7 +60,7 @@
                     {{-- "Three Dots" Separator --}}
                     @if (is_string($element))
                         <span aria-disabled="true">
-                            <button type="button" class="btn px-1 py-05" disabled>{{ $element }}</button>
+                            <button type="button" class="btn wh-2.5" disabled>{{ $element }}</button>
                         </span>
                     @endif
 
@@ -69,11 +69,11 @@
                         @foreach ($element as $page => $url)
                             <span wire:key="paginator-{{ $paginator->getPageName() }}-page{{ $page }}">
                                 @if ($page == $paginator->currentPage())
-                                    <button type="button" class="btn px-1 py-05" disabled>{{ $page }}</button>
+                                    <button type="button" class="btn wh-2.5" disabled>{{ $page }}</button>
                                 @else
                                     <button wire:click="gotoPage({{ $page }}, '{{ $paginator->getPageName() }}')"
                                         x-on:click="{{ $scrollIntoViewJsSnippet }}" aria-label="Go to page {{ $page }}"
-                                        type="button" class="btn px-1 py-05"> {{ $page }} </button>
+                                        type="button" class="btn wh-2.5"> {{ $page }} </button>
                                 @endif
                             </span>
                         @endforeach
@@ -84,13 +84,13 @@
                 <div class="inline-flex">
                     @if ($paginator->hasMorePages())
                         <button wire:click="nextPage('{{ $paginator->getPageName() }}')" x-on:click="{{ $scrollIntoViewJsSnippet }}"
-                            type="button" class="btn px-1 py-05" aria-label="Next page">
+                            type="button" class="btn wh-2.5" aria-label="Next page">
                             <svg class="icon" fill="currentColor" viewBox="0 0 20 20">
                                 <path fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clip-rule="evenodd" />
                             </svg>
                         </button>
                     @else
-                        <button type="button" class="btn px-1 py-05" disabled aria-label="Next page">
+                        <button type="button" class="btn wh-2.5" disabled aria-label="Next page">
                             <svg class="icon txt-muted" fill="currentColor" viewBox="0 0 20 20">
                                 <path fill-rule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clip-rule="evenodd" />
                             </svg>